wip: preserve foreign settings-groups refactor found in this worktree
Not part of #3318 (already merged) -- this content bled in from another
process/job per a prior session's handover note. Preserving ahead of a
machine migration rather than losing it; ownership/purpose unconfirmed.
